// intellishop-frontend/src/app/api/generated/model/authResponse.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AuthResponse { 
    token?: string;
    userId?: number;
    email?: string;
    role?: string;
}